Resolve conflicts across 45 files, reconciling the zig 0.16 migration
(zig-16 branch) with origin/main's refactors (which still target 0.15.2).

Resolution principle: main's logic/architecture wins; zig-16's 0.16 API
adaptations win; merge both where they overlap.

Notable decisions:
- Accept main's deletion of the MIR/corecir pipeline and builtins/handlers.zig
  (signal handling now via base/signal_handler.zig).
- Adopt main's LirImage execution path in the CLI (viewLirImageFromHandle,
  evaluateLirImageEntrypoint), keeping zig-16's CliCtx/ctx.io.std_io and
  EchoEnv conventions.
- Restore main's semantic_audit.pl + SemanticAuditStep (ported to the 0.16
  std.process.spawn pattern) and its minici wiring.
- Keep zig-16's std.Io threading, std.Io.File, std.process.run/spawn,
  unmanaged ArrayList, find/findScalar, and std.c.write usage throughout.

Migration of origin/main's new (0.15.2) code to 0.16 is ongoing; the tree
is not yet expected to fully build.

+++ src/canonicalize/TypeAnnotation.zig
@@ -345,7 +345,7 @@ pub const TypeAnno = union(enum) {
         },
         external: struct {
             module_idx: CIR.Import.Idx,
-            target_node_idx: u16,
+            target_node_idx: u32,
         },
         /// Pending external lookup - deferred until dependencies are canonicalized
         pending: struct {
